Changed dgen file structure
